updated MR interaction parameters; consistent real kinds (wp), change to lower case Fortran keywords
Fixed an error in the value for the CHnO[ether] <-–> H+ interaction parameter, which was introduced with v3.0 to v3.05. Thanks to Florian Couvidat for reporting this issue. Due to the analogy approached used, the error in the mentioned parameter also affected the tabulated values for H+ with the (CHnOOH[perox]) and (CHnOOCHm[perox]) main groups. Those tabulated values were then updated alongside with other code for AIOMFAC-web v3.0. However, the older values were correct. 
The changes with this code update also include some general notation changes of Fortran keywords to lower case, a few updates to SR interaction parameters via the analogy approach to cover a selection of nitrogen-containing groups that previously did not have set interaction parameters.