updated SR and MR interaction parameter arrays
updated SR and MR interaction parameter arrays, specifically ARR short-range interaction parameters related to a subset of main group interactions involving the nitrogen-containing groups CNO2 and ACNO2 as well as ACOH and the oxyethylene group CH2OCH2[PEG]. For the latter, an analogy approach was used to set unknown interaction parameters in ARR as well as for the middle-range interactions with inorganic ions. Future improvements will be possible with proper fitting of pertaining parameters using reliable experimental data.